Adding `jax.Array` to jax.experimental. Its pretty much the same as GDA (without the performance optimization for now).
Currently, jax.Array takes DeviceArrays in `assemble_array` because device_put returns a DA. In the future (with IFRT), it will return an `Array`.

`explode_into_addressable_arrays` wraps DA into jax.Array with a `SingleDeviceSharding`. `assemble_array` is the counterpart to `explode_into_addressable_arrays` which takes in arrays with single device sharding (currently DA) and creates an `Array` from it.

PiperOrigin-RevId: 452088043